Moved tweet model to a separate object
