+++ src/main/scala/com/alawresz/twitterment/web/AkkaSystem.scala
@@ -2,8 +2,9 @@ package com.alawresz.twitterment.web
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
+import com.typesafe.scalalogging.LazyLogging
 
-trait AkkaSystem {
+trait AkkaSystem extends LazyLogging {
   implicit val system       = ActorSystem()
   implicit val materializer = ActorMaterializer()
